+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.46.17](https://github.com/adobe/spacecat-audit-worker/compare/v1.46.16...v1.46.17) (2025-01-15)
+
+
+### Bug Fixes
+
+* removed sendMessage for broken-backlinks ([#587](https://github.com/adobe/spacecat-audit-worker/issues/587)) ([166d209](https://github.com/adobe/spacecat-audit-worker/commit/166d2097cd38cc7f8601e4bd156e0f426bebc316))
+
+## [1.46.16](https://github.com/adobe/spacecat-audit-worker/compare/v1.46.15...v1.46.16) (2025-01-15)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-utils to v1.26.3 ([#584](https://github.com/adobe/spacecat-audit-worker/issues/584)) ([4bbc51c](https://github.com/adobe/spacecat-audit-worker/commit/4bbc51ca9dbe4778899396d6778a259407e7b954))
+
 ## [1.46.15](https://github.com/adobe/spacecat-audit-worker/compare/v1.46.14...v1.46.15) (2025-01-13)